pong-jorge u
Learning how to do line art with JavaScript for the first time, I kept it basic with Pong to get the hang of things. Since Pong is easily recognizable, I felt that it made for a good starting point. I definitely want to learn how to create more detailed patterns like the ones in the gallery.