Uninterrupted
This scene is meant to describe the beauty of nature through code. It is also meant to describe how Uninterrupted and peaceful you feel when in nature. I have a background in coding, but that is just in Python. This was a new experience and I enjoyed it. I worked on this for some time, so I hope it gets in!