Significantly improve error messages
This adds assertations to all of the built-in
functions in Blot to throw meaningful errors when
the wrong arguments are passed. It also upgrades
the stack trace parsing logic, using V8's object
representation and a regex built based on the
SpiderMonkey and JavaScriptCore source code.
Safari (JavaScriptCore) does not keep track of
line numbers within eval'd code, so a hack is used
to construct a data URL and import it when it is
detected to be nessesary.

The previous version of Blot would give unhelpful
errors like "t is not defined" on line 0, and line
numbers only ever worked on Chrome in some cases.
Now, it will give helpful errors explaining what
went wrong, pointing to the correct position in
the code on any browser.

There is a known issue where, on Safari, the stack
does not follow nesting, so if the user defines a
function, the error will appear where the function
is called instead of within the function, but
there is no way to fix this.

This commit is a combination of these 9 commits:

Add code to assert the type of an arguments array
Refactor assertion code and begin adding it to functions
Add assertations to the rest of the Blot functions
Fix bugs and make the assertations work
Begin work on properly parsing error stack traces
Extract SpiderMonkey (Firefox) stack trace position correctly
Work on making the stack trace extraction function on Safari
Make stack traces work on Safari
Activate Safari hack when Safari stack traces detected