Add files via upload
Applied changes suggested by @BrightTheBackpack! I added a second flower, randomized the number of leaves (if random seed is even, it shows the right leaf, if divisible by 3, it shows the left one, and if it's greater than 3, it shows the leaf on the second flower), and added a ground. Thanks for the feedback!